==> 3-GatPrediction.ipynb <==
#Get the model
path_result='C:\\Backup of covid project\\2cls_negVSnonneg\\results\\'
path_data="C:\\Backup of covid project\\2cls_negVSnonneg\\data\\"
path_code="C:\\Backup of covid project\\"

import time
import numpy as np
import tensorflow as tf
import os
import random
os.chdir(path_code)
import gat
import imp
imp.reload(gat)
from gat import GAT, HeteGAT, HeteGAT_multi  # or * for that matter
import process
import importlib
importlib.reload(process)
import pandas as pd
import pickle
import scipy.sparse
from scipy.sparse import csc_matrix
from scipy import sparse
import csv 

os.environ["CUDA_VISIBLE_DEVICES"] = "1,2,3"

config = tf.ConfigProto()
config.gpu_options.allow_growth = True

dataset = 'acm'
featype = 'fea'
checkpt_file=path_result+'acm_allMP_multi_fea_.ckpt'

print('model: {}'.format(checkpt_file))
# training params for only CC
batch_size = 1
nb_epochs = 100
patience = 100
lr = 0.02  # learning rate
l2_coef = 0.08#0.1#0.001  # weight decay
# numbers of hidden units per each attention head in each layer
hid_units = [8]
n_heads = [4, 1]  # additional entry for the output layer
residual = False
nonlinearity = tf.nn.elu
model = HeteGAT_multi

print('Dataset: ' + dataset)
print('----- Opt. hyperparams -----')
print('lr: ' + str(lr))
print('l2_coef: ' + str(l2_coef))
print('----- Archi. hyperparams -----')
print('nb. layers: ' + str(len(hid_units)))
print('nb. units per layer: ' + str(hid_units))
print('nb. attention heads: ' + str(n_heads))
print('residual: ' + str(residual))
print('nonlinearity: ' + str(nonlinearity))
print('model: ' + str(model))

# jhy data
import scipy.io as sio
import scipy.sparse as sp


def sample_mask(idx, l):
    """Create mask."""
    mask = np.zeros(l)
    mask[idx] = 1
    return np.array(mask, dtype=np.bool)

def load_data_dblp(path=path_data):
    schools = ["notredame","uofm","columbia","dartmouth","UCSD","berkeley","Harvard","ucla"]
    s=schools[3]#use this index to control which network we are going to run.
    t="comment"
    
    os.chdir(path)
    name="label_emt3_"+s+"_cm.csv"
    truelabels=pd.read_csv(name,skip_blank_lines=True,header=None)
    name="feature_"+s+t+".csv"
    truefeatures=pd.read_csv(name,skip_blank_lines=True,header=None).values

    N=truefeatures.shape[0]
    name="CCsym_"+s+".npz"
    dat_cc=scipy.sparse.load_npz(name);dat_cc=dat_cc.toarray();
    rownetworks = [(dat_cc)]

    y=truelabels
    name="train_index_"+s+"_cm.p"
    train_idx = pickle.load(open(name,"rb"));train_idx=np.array(train_idx)
    name="to_be_labeled_index_"+s+"_cm.p"
    val_idx = pickle.load(open(name,"rb"));val_idx=np.array(val_idx)
    name="test_index_"+s+"_cm.p"
    test_idx = pickle.load(open(name,"rb"));test_idx=np.array(test_idx)

    train_mask = sample_mask(train_idx, y.shape[0])
    val_mask = sample_mask(val_idx, y.shape[0])
    test_mask = sample_mask(test_idx, y.shape[0])
    
    y_train = np.zeros(y.shape)
    y_val = np.zeros(y.shape)
    y_test = np.zeros(y.shape)
    y=y.values
    y_train[train_mask, :] = y[train_mask, :]
    y_val[val_mask, :] = y[val_mask, :]
    y_test[test_mask, :] = y[test_mask, :]

    # return selected_idx, selected_idx_2
    print('y_train:{}, y_val:{}, y_test:{}, train_idx:{}, val_idx:{}, test_idx:{}'.format(y_train.shape,
                                                                                          y_val.shape,
                                                                                          y_test.shape,
                                                                                          train_idx.shape,
                                                                                          val_idx.shape,
                                                                                          test_idx.shape))
    truefeatures_list = [truefeatures, truefeatures]#, truefeatures, truefeatures]# ?? why copy three times? First for center node, 2 for each metapath.    
    os.chdir(path_code)
    return rownetworks, truefeatures_list, y_train, y_val, y_test, train_mask, val_mask, test_mask


# use adj_list as fea_list, have a try~
adj_list, fea_list, y_train, y_val, y_test, train_mask, val_mask, test_mask = load_data_dblp()
if featype == 'adj':
    fea_list = adj_list

import scipy.sparse as sp

nb_nodes = fea_list[0].shape[0]
ft_size = fea_list[0].shape[1]
nb_classes = y_train.shape[1]

fea_list = [fea[np.newaxis] for fea in fea_list]
adj_list = [adj[np.newaxis] for adj in adj_list]
y_train = y_train[np.newaxis]
y_val = y_val[np.newaxis]
y_test = y_test[np.newaxis]
train_mask = train_mask[np.newaxis]
val_mask = val_mask[np.newaxis]
test_mask = test_mask[np.newaxis]

biases_list = [process.adj_to_bias(adj, [nb_nodes], nhood=1) for adj in adj_list]
#biases_list = [process.adj_to_bias(adj) for adj in adj_list]
#process.adj_to_bias(adj)
print('build graph...')
with tf.Graph().as_default():
    with tf.name_scope('input'):
        ftr_in_list = [tf.placeholder(dtype=tf.float32,
                                      shape=(batch_size, nb_nodes, ft_size),
                                      name='ftr_in_{}'.format(i))
                       for i in range(len(fea_list))]
        bias_in_list = [tf.placeholder(dtype=tf.float32,
                                       shape=(batch_size, nb_nodes, nb_nodes),
                                       name='bias_in_{}'.format(i))
                        for i in range(len(biases_list))]
        lbl_in = tf.placeholder(dtype=tf.int32, shape=(
            batch_size, nb_nodes, nb_classes), name='lbl_in')
        msk_in = tf.placeholder(dtype=tf.int32, shape=(batch_size, nb_nodes),
                                name='msk_in')
        attn_drop = tf.placeholder(dtype=tf.float32, shape=(), name='attn_drop')
        ffd_drop = tf.placeholder(dtype=tf.float32, shape=(), name='ffd_drop')
        is_train = tf.placeholder(dtype=tf.bool, shape=(), name='is_train')
    # forward
    logits, final_embedding, att_val = model.inference(ftr_in_list, nb_classes, nb_nodes, is_train,
                                                       attn_drop, ffd_drop,
                                                       bias_mat_list=bias_in_list,
                                                       hid_units=hid_units, n_heads=n_heads,
                                                       residual=residual, activation=nonlinearity)

    # cal masked_loss
    log_resh = tf.reshape(logits, [-1, nb_classes])
    lab_resh = tf.reshape(lbl_in, [-1, nb_classes])
    msk_resh = tf.reshape(msk_in, [-1])
    loss = model.masked_softmax_cross_entropy(log_resh, lab_resh, msk_resh)
    predicted=tf.argmax(log_resh, 1)
    accuracy = model.masked_accuracy(log_resh, lab_resh, msk_resh)
    # optimzie
    train_op = model.training(loss, lr, l2_coef)

    saver = tf.train.Saver()

    init_op = tf.group(tf.global_variables_initializer(),
                       tf.local_variables_initializer())

    vlss_mn = np.inf
    vacc_mx = 0.0
    curr_step = 0

    with tf.Session(config=config) as sess:
        sess.run(init_op)
        train_loss_avg = 0
        train_acc_avg = 0
        val_loss_avg = 0
        val_acc_avg = 0
        val_pred_avg = 0
        for epoch in range(nb_epochs):
            tr_step = 0
           
            tr_size = fea_list[0].shape[0]
            # ================   training    ============
            while tr_step * batch_size < tr_size:

                fd1 = {i: d[tr_step * batch_size:(tr_step + 1) * batch_size]
                       for i, d in zip(ftr_in_list, fea_list)}
                fd2 = {i: d[tr_step * batch_size:(tr_step + 1) * batch_size]
                       for i, d in zip(bias_in_list, biases_list)}
                fd3 = {lbl_in: y_train[tr_step * batch_size:(tr_step + 1) * batch_size],
                       msk_in: train_mask[tr_step * batch_size:(tr_step + 1) * batch_size],
                       is_train: True,
                       attn_drop: 0,
                       ffd_drop: 0}
                fd = fd1
                fd.update(fd2)
                fd.update(fd3)
                _, loss_value_tr, acc_tr, att_val_train = sess.run([train_op, loss, accuracy, att_val],
                                                                   feed_dict=fd)
                train_loss_avg += loss_value_tr
                train_acc_avg += acc_tr
                tr_step += 1

            vl_step = 0
            vl_size = fea_list[0].shape[0]

            # import pdb; pdb.set_trace()
            print('Epoch: {}, att_val: {}'.format(epoch, np.mean(att_val_train, axis=0)))
            print('Training: loss = %.5f, acc = %.5f |' %
                  (train_loss_avg / tr_step, train_acc_avg / tr_step))

            curr_step += 1
            if epoch==nb_epochs-1:
                saver.save(sess, checkpt_file)

            train_loss_avg = 0
            train_acc_avg = 0
            val_loss_avg = 0
            val_acc_avg = 0
        saver.restore(sess, checkpt_file)
        print('load model from : {}'.format(checkpt_file))
        ts_size = fea_list[0].shape[0]
        ts_step = 0
        ts_loss = 0.0
        ts_acc = 0.0

        while ts_step * batch_size < ts_size:
            fd1 = {i: d[ts_step * batch_size:(ts_step + 1) * batch_size]
                   for i, d in zip(ftr_in_list, fea_list)}
            fd2 = {i: d[ts_step * batch_size:(ts_step + 1) * batch_size]
                   for i, d in zip(bias_in_list, biases_list)}
            fd3 = {lbl_in: y_test[ts_step * batch_size:(ts_step + 1) * batch_size],
                   msk_in: test_mask[ts_step * batch_size:(ts_step + 1) * batch_size],
            
                   is_train: False,
                   attn_drop: 0.0,
                   ffd_drop: 0.0}
        
            fd = fd1
            fd.update(fd2)
            fd.update(fd3)
            loss_value_ts, acc_ts, jhy_final_embedding = sess.run([loss, accuracy, final_embedding],
                                                                  feed_dict=fd)
            ts_loss += loss_value_ts
            ts_acc += acc_ts
            ts_step += 1

        print('Test loss:', ts_loss / ts_step,
              '; Test accuracy:', ts_acc / ts_step)
        
        print('start knn, kmean.....')
        xx = np.expand_dims(jhy_final_embedding, axis=0)[test_mask]

        from numpy import linalg as LA

        yy = y_test[test_mask]

        print('xx: {}, yy: {}'.format(xx.shape, yy.shape))
        from jhyexp import my_KNN, my_Kmeans#, my_TSNE, my_Linear

        my_KNN(xx, yy)
        my_Kmeans(xx, yy)

        sess.close()


#run previous block before running this one
#get the testing result without submission data, i.e., the graph that this block used only contains comments data
#the printed output is meaningless
#predicted results are saved to .p file
#prediction in this step is used in step 4, but not in step 5
#if you have memory error, calculate "biases_list" variable for each school and save it locally, then jsut load during running. The code for this is commented out and you can use it directly


import time
import numpy as np
import tensorflow as tf
import os
os.chdir(path_code)
from gat import GAT, HeteGAT, HeteGAT_multi
import process
import importlib
from layers import attn_head, SimpleAttLayer
importlib.reload(process)
import pandas as pd
import pickle
import scipy.sparse
from scipy.sparse import csc_matrix
import scipy import sparse
import scipy.io as sio
import scipy.sparse as sp

import csv 
config = tf.ConfigProto()
config.gpu_options.allow_growth = True

dataset = 'acm'
featype = 'fea'
print('Dataset: ' + dataset)
print('----- Opt. hyperparams -----')

# jhy data
import scipy.io as sio
import scipy.sparse as sp


def sample_mask(idx, l):
    """Create mask."""
    mask = np.zeros(l)
    mask[idx] = 1
    return np.array(mask, dtype=np.bool)

schools = ["notredame2019","notredame2020","uofm2019","uofm2020","columbia2019","columbia2020","dartmouth","UCSD2019","UCSD2020","berkeley2019","berkeley2020","Harvard2019","Harvard2020","ucla2019","ucla2020"]
schools_type = ["notredamecomment2019","notredamecomment2020","uofmcomment2019","uofmcomment2020","columbiacomment2019","columbiacomment2020","dartmouthcomment","UCSDcomment2019","UCSDcomment2020",
           "berkeleycomment2019","berkeleycomment2020","Harvardcomment2019","Harvardcomment2020","uclacomment2019","uclacomment2020"]

def load_data_dblp(s,st,path=path_data):
    t="comment"
    
    os.chdir(path)
    name="label_emt3_"+s+"_cm.csv"
    truelabels=pd.read_csv(name,skip_blank_lines=True,header=None)
    name="feature_"+st+".csv"
    truefeatures=pd.read_csv(name,skip_blank_lines=True,header=None).values

    N=truefeatures.shape[0]
    name="CCsym_"+s+".npz"
    dat_cc=scipy.sparse.load_npz(name);dat_cc=dat_cc.toarray();
    rownetworks = [(dat_cc)]

    y=truelabels
    name="train_index_"+s+"_cm.p"
    train_idx = pickle.load(open(name,"rb"));train_idx=np.array(train_idx)
    name="to_be_labeled_index_"+s+"_cm.p"
    val_idx = pickle.load(open(name,"rb"));val_idx=np.array(val_idx)
    name="test_index_"+s+"_cm.p"
    test_idx = pickle.load(open(name,"rb"));test_idx=np.array(test_idx)
    print(len(test_idx))
    #use this line to control testing set
    test_idx=np.array(range(N))
    
    train_mask = sample_mask(train_idx, y.shape[0])
    val_mask = sample_mask(val_idx, y.shape[0])
    test_mask = sample_mask(test_idx, y.shape[0])# ?? why copy three times? First for center node, 2 for each metapath.
    
    y_train = np.zeros(y.shape)
    y_val = np.zeros(y.shape)
    y_test = np.zeros(y.shape)
    y=y.values
    y_train[train_mask, :] = y[train_mask, :]
    y_val[val_mask, :] = y[val_mask, :]
    y_test[test_mask, :] = y[test_mask, :]

    # return selected_idx, selected_idx_2
    print('y_train:{}, y_val:{}, y_test:{}, train_idx:{}, val_idx:{}, test_idx:{}'.format(y_train.shape,
                                                                                          y_val.shape,
                                                                                          y_test.shape,
                                                                                          train_idx.shape,
                                                                                          val_idx.shape,
                                                                                          test_idx.shape))
    truefeatures_list = [truefeatures, truefeatures]#, truefeatures, truefeatures]# ?? why copy three times? First for center node, 2 for each metapath.    
    os.chdir(path_code)
    return rownetworks, truefeatures_list, y_train, y_val, y_test, train_mask, val_mask, test_mask



for i in range(len(schools)):
    adj_list, fea_list, y_train, y_val, y_test, train_mask, val_mask, test_mask = load_data_dblp(s=schools[i],st=schools_type[i])
    

    if featype == 'adj':
        fea_list = adj_list



    import scipy.sparse as sp

    nb_nodes = fea_list[0].shape[0]
    ft_size = fea_list[0].shape[1]
    nb_classes = y_train.shape[1]
    
    fea_list = [fea[np.newaxis] for fea in fea_list]
    adj_list = [adj[np.newaxis] for adj in adj_list]
    y_train = y_train[np.newaxis]
    y_val = y_val[np.newaxis]
    y_test = y_test[np.newaxis]
    train_mask = train_mask[np.newaxis]
    val_mask = val_mask[np.newaxis]
    test_mask = test_mask[np.newaxis]
    print("s")
    
    biases_list = [process.adj_to_bias(adj, [nb_nodes], nhood=1) for adj in adj_list]
    name=schools[i]+'adj.p'
    os.chdir(path_data)
    pickle.dump(biases_list,open(name,"wb"))
    #biases_list=pickle.load(open(name,"rb"))
    
    
    print('build graph...')
    with tf.Graph().as_default():
        with tf.name_scope('input'):
            ftr_in_list = [tf.placeholder(dtype=tf.float32,
                                          shape=(batch_size, nb_nodes, ft_size),
                                          name='ftr_in_{}'.format(i))
                           for i in range(len(fea_list))]
            bias_in_list = [tf.placeholder(dtype=tf.float32,
                                           shape=(batch_size, nb_nodes, nb_nodes),
                                           name='bias_in_{}'.format(i))
                            for i in range(len(biases_list))]
            lbl_in = tf.placeholder(dtype=tf.int32, shape=(
                batch_size, nb_nodes, nb_classes), name='lbl_in')
            msk_in = tf.placeholder(dtype=tf.int32, shape=(batch_size, nb_nodes),
                                    name='msk_in')
            attn_drop = tf.placeholder(dtype=tf.float32, shape=(), name='attn_drop')
            ffd_drop = tf.placeholder(dtype=tf.float32, shape=(), name='ffd_drop')
            is_train = tf.placeholder(dtype=tf.bool, shape=(), name='is_train')
        # forward
        logits, final_embedding, att_val = model.inference(ftr_in_list, nb_classes, nb_nodes, is_train,attn_drop, ffd_drop,bias_mat_list=bias_in_list,hid_units=hid_units, n_heads=n_heads,residual=residual, activation=nonlinearity)
    
        # cal masked_loss
        log_resh = tf.reshape(logits, [-1, nb_classes])
        lab_resh = tf.reshape(lbl_in, [-1, nb_classes])
        msk_resh = tf.reshape(msk_in, [-1])
        loss = model.masked_softmax_cross_entropy(log_resh, lab_resh, msk_resh)
        accuracy = model.masked_accuracy(log_resh, lab_resh, msk_resh)
        # optimzie
        train_op = model.training(loss, lr, l2_coef)
        vlss_mn = np.inf
        vacc_mx = 0.0
        curr_step = 0
        saver = tf.train.Saver()
        init_op = tf.group(tf.global_variables_initializer(),
                           tf.local_variables_initializer())
        with tf.Session(config=config) as sess:
            #sess.run(init_op)

            train_loss_avg = 0
            train_acc_avg = 0
            val_loss_avg = 0
            val_acc_avg = 0
            os.chdir(path_code)
            saver.restore(sess, path_result+"acm_allMP_multi_fea_.ckpt")
            print("Model restored.")
            
            ts_size = fea_list[0].shape[0]
            print(fea_list[0].shape)
            ts_step = 0
            ts_loss = 0.0
            ts_acc = 0.0
            print("test start")
            while ts_step * batch_size < ts_size:
                print(ts_step)
                print(batch_size)
                print(ts_size)
                # fd1 = {ftr_in: features[ts_step * batch_size:(ts_step + 1) * batch_size]}
                fd1 = {i: d[ts_step * batch_size:(ts_step + 1) * batch_size]
                       for i, d in zip(ftr_in_list, fea_list)}
                fd2 = {i: d[ts_step * batch_size:(ts_step + 1) * batch_size]
                       for i, d in zip(bias_in_list, biases_list)}
                fd3 = {lbl_in: y_test[ts_step * batch_size:(ts_step + 1) * batch_size],
                       msk_in: test_mask[ts_step * batch_size:(ts_step + 1) * batch_size],
                
                       is_train: False,
                       attn_drop: 0.0,
                       ffd_drop: 0.0}
            
                fd = fd1
                fd.update(fd2)
                fd.update(fd3)
                
                loss_value_ts, acc_ts, jhy_final_embedding, predicted = sess.run([loss, accuracy, final_embedding, log_resh],
                                                                      feed_dict=fd)
                import pickle
                name=schools[i]+'.p'
                os.chdir(path_result)
                pickle.dump(predicted,open(name,"wb"))
                ts_loss += loss_value_ts
                ts_acc += acc_ts
                ts_step += 1

            print('Test loss:', ts_loss / ts_step,
                  '; Test accuracy:', ts_acc / ts_step)
    
            
            sess.close()


#run first block before running this one
#get the testing result WITH submission data, i.e., the graph that this block contains comments and submission
#the printed output is meaningless
#predicted results are saved to .p file
#prediction in this step is used in step 5, but not in step 4
#if you have memory error, calculate "biases_list" variable for each school and save it locally. The code for this is commented out and you can use it directly

#The feature, index, adjacency, label data are all different from the one used in previou block, contact the author to get these two sets of data

path_result='/content/drive/MyDrive/Colab Notebooks/results/'
path_data="/content/drive/MyDrive/Colab Notebooks/data/"
path_code="/content/drive/MyDrive/Colab Notebooks/"

import time
import numpy as np
import tensorflow as tf
import os
os.chdir(path_code)
from gat import GAT, HeteGAT, HeteGAT_multi
import process
import importlib
from layers import attn_head, SimpleAttLayer
importlib.reload(process)
import pandas as pd
import pickle
import scipy.sparse
from scipy.sparse import csc_matrix
import scipy import sparse
import scipy.io as sio
import scipy.sparse as sp

import csv 
    
os.environ["CUDA_VISIBLE_DEVICES"] = "1,2,3"

config = tf.ConfigProto()
config.gpu_options.allow_growth = True

dataset = 'acm'
featype = 'fea'
# training params

checkpt_file=path_result+'acm_allMP_multi_fea_.ckpt'


print('Dataset: ' + dataset)
print('----- Opt. hyperparams -----')



def sample_mask(idx, l):
    """Create mask."""
    mask = np.zeros(l)
    mask[idx] = 1
    return np.array(mask, dtype=np.bool)

#schools = ["berkeley2020"]
schools =["notredame2019","notredame2020","uofm2019","uofm2020","columbia2019","columbia2020","dartmouth","UCSD2019","UCSD2020","berkeley2019","berkeley2020","Harvard2019","Harvard2020","ucla2019","ucla2020"]
schools_type = ["notredamecomment2019","notredamecomment2020","uofmcomment2019","uofmcomment2020","columbiacomment2019","columbiacomment2020","dartmouthcomment","UCSDcomment2019","UCSDcomment2020",
           "berkeleycomment2019","berkeleycomment2020","Harvardcomment2019","Harvardcomment2020","uclacomment2019","uclacomment2020"]


#path_data='C:\\data\\HAN-modified\\tranfer learning\\'
def load_data_dblp(s,path=path_data):
    t="comment"
    
    os.chdir(path)
    name="label_emt3_"+s+".csv"
    truelabels=pd.read_csv(name,skip_blank_lines=True,header=None)
    name="feature_"+s+".csv"
    truefeatures=pd.read_csv(name,skip_blank_lines=True,header=None).values
    N=truefeatures.shape[0]
    name="CCsym_"+s+".npz"
    dat_cc=scipy.sparse.load_npz(name);
    dat_cc=dat_cc.toarray();
    rownetworks = [(dat_cc)]

    y=truelabels
    name="train_index_"+s+"_cm.p"
    train_idx = pickle.load(open(name,"rb"));train_idx=np.array(train_idx)
    name="to_be_labeled_index_"+s+"_cm.p"
    val_idx = pickle.load(open(name,"rb"));val_idx=np.array(val_idx)
    name="test_index_"+s+"_cm.p"
    test_idx = pickle.load(open(name,"rb"));test_idx=np.array(test_idx)
    test_idx=np.array(range(N))
    print(N)
    
    train_mask = sample_mask(train_idx, y.shape[0])
    val_mask = sample_mask(val_idx, y.shape[0])
    test_mask = sample_mask(test_idx, y.shape[0])# ?? why copy three times? First for center node, 2 for each metapath.
    
    y_train = np.zeros(y.shape)
    y_val = np.zeros(y.shape)
    y_test = np.zeros(y.shape)
    y=y.values
    y_train[train_mask, :] = y[train_mask, :]
    y_val[val_mask, :] = y[val_mask, :]
    y_test[test_mask, :] = y[test_mask, :]

    # return selected_idx, selected_idx_2
    print('y_train:{}, y_val:{}, y_test:{}, train_idx:{}, val_idx:{}, test_idx:{}'.format(y_train.shape,
                                                                                          y_val.shape,
                                                                                          y_test.shape,
                                                                                          train_idx.shape,
                                                                                          val_idx.shape,
                                                                                          test_idx.shape))
    truefeatures_list = [truefeatures, truefeatures]#, truefeatures, truefeatures]# ?? why copy three times? First for center node, 2 for each metapath.    
    os.chdir(path_code)
    return rownetworks, truefeatures_list, y_train, y_val, y_test, train_mask, val_mask, test_mask



for i in range(len(schools)):
    adj_list, fea_list, y_train, y_val, y_test, train_mask, val_mask, test_mask = load_data_dblp(s=schools[i])
    if featype == 'adj':
        fea_list = adj_list

    nb_nodes = fea_list[0].shape[0]
    ft_size = fea_list[0].shape[1]
    nb_classes = y_train.shape[1]

    fea_list = [fea[np.newaxis] for fea in fea_list]
    adj_list = [adj[np.newaxis] for adj in adj_list]
    y_train = y_train[np.newaxis]
    y_val = y_val[np.newaxis]
    y_test = y_test[np.newaxis]
    train_mask = train_mask[np.newaxis]
    val_mask = val_mask[np.newaxis]
    test_mask = test_mask[np.newaxis]
    print("s")
    
    biases_list = [process.adj_to_bias(adj, [nb_nodes], nhood=1) for adj in adj_list]
    import pickle
    name=schools[i]+'adj.p'
    os.chdir(path_data)
    pickle.dump(biases_list,open(name,"wb"))
    #biases_list=pickle.load(open(name,"rb"))
    
    print('build graph...')
    with tf.Graph().as_default():
        with tf.name_scope('input'):
            ftr_in_list = [tf.placeholder(dtype=tf.float32,
                                          shape=(batch_size, nb_nodes, ft_size),
                                          name='ftr_in_{}'.format(i))
                           for i in range(len(fea_list))]
            bias_in_list = [tf.placeholder(dtype=tf.float32,
                                           shape=(batch_size, nb_nodes, nb_nodes),
                                           name='bias_in_{}'.format(i))
                            for i in range(len(biases_list))]
            lbl_in = tf.placeholder(dtype=tf.int32, shape=(
                batch_size, nb_nodes, nb_classes), name='lbl_in')
            msk_in = tf.placeholder(dtype=tf.int32, shape=(batch_size, nb_nodes),
                                    name='msk_in')
            attn_drop = tf.placeholder(dtype=tf.float32, shape=(), name='attn_drop')
            ffd_drop = tf.placeholder(dtype=tf.float32, shape=(), name='ffd_drop')
            is_train = tf.placeholder(dtype=tf.bool, shape=(), name='is_train')
        # forward
        logits, final_embedding, att_val = model.inference(ftr_in_list, nb_classes, nb_nodes, is_train,attn_drop, ffd_drop,bias_mat_list=bias_in_list,hid_units=hid_units, n_heads=n_heads,residual=residual, activation=nonlinearity)
    
        # cal masked_loss
        log_resh = tf.reshape(logits, [-1, nb_classes])
        lab_resh = tf.reshape(lbl_in, [-1, nb_classes])
        msk_resh = tf.reshape(msk_in, [-1])
        loss = model.masked_softmax_cross_entropy(log_resh, lab_resh, msk_resh)
        accuracy = model.masked_accuracy(log_resh, lab_resh, msk_resh)
        # optimzie
        train_op = model.training(loss, lr, l2_coef)
        vlss_mn = np.inf
        vacc_mx = 0.0
        curr_step = 0
        saver = tf.train.Saver()
        init_op = tf.group(tf.global_variables_initializer(),
                           tf.local_variables_initializer())
        with tf.Session(config=config) as sess:
            #sess.run(init_op)

            train_loss_avg = 0
            train_acc_avg = 0
            val_loss_avg = 0
            val_acc_avg = 0
            os.chdir(path_code)
            saver.restore(sess, path_result+"acm_allMP_multi_fea_.ckpt")
            print("Model restored.")
            
            ts_size = fea_list[0].shape[0]
            print(fea_list[0].shape)
            ts_step = 0
            ts_loss = 0.0
            ts_acc = 0.0
            print("test start")
            while ts_step * batch_size < ts_size:
                print(ts_step)
                print(batch_size)
                print(ts_size)
                # fd1 = {ftr_in: features[ts_step * batch_size:(ts_step + 1) * batch_size]}
                fd1 = {i: d[ts_step * batch_size:(ts_step + 1) * batch_size]
                       for i, d in zip(ftr_in_list, fea_list)}
                fd2 = {i: d[ts_step * batch_size:(ts_step + 1) * batch_size]
                       for i, d in zip(bias_in_list, biases_list)}
                fd3 = {lbl_in: y_test[ts_step * batch_size:(ts_step + 1) * batch_size],
                       msk_in: test_mask[ts_step * batch_size:(ts_step + 1) * batch_size],
                
                       is_train: False,
                       attn_drop: 0.0,
                       ffd_drop: 0.0}
                import psutil
              
                print(psutil.virtual_memory().available * 100 / psutil.virtual_memory().total)
                fd = fd1
                fd.update(fd2)
                fd.update(fd3)
                print('fini')
                del bias_in_list
                del biases_list
                del ftr_in_list
                del fea_list
                print('dele finished')
                print(psutil.virtual_memory().available * 100 / psutil.virtual_memory().total)
                loss_value_ts, acc_ts, jhy_final_embedding, predicted = sess.run([loss, accuracy, final_embedding, log_resh],
                                                                      feed_dict=fd)
                import pickle
                name=schools[i]+'.p'
                os.chdir(path_result)
                pickle.dump(predicted,open(name,"wb"))
                ts_loss += loss_value_ts
                ts_acc += acc_ts
                ts_step += 1

            print('Test loss:', ts_loss / ts_step,
                  '; Test accuracy:', ts_acc / ts_step)
    
            
            sess.close()
